add release number
the idea of the release number is to be incremented more often than
build, everytime a new feature is added
this is helpful in development, because during dev, the build number
will stay at the previous released build number until the full release
is done. that means any <BuildGuards /> will trigger until then, making
testing very annoying
<BuildGuards /> will be extended to allow a minimum release number too
thats the concept, anyway. watch me forget about this in a week's time